fix tests and remove duplicate code
